requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
